Add_Track_Popup Functionality Test Push
Set the player to be its own scene and refactored the code so that it could be instantiated in the main window and pull double duty as the track player. Changed the song progress tracker to a scroll bar instead of a value bar, and shifted the value bar to control volume. Added the ability to numerically set the timestamps for tags, to allow for more precise adjustments. Will need to do some user testing to make sure interface is functional before moving on to core functionality